refactor(kotlin): ドメインを 3 層パッケージへ再編成し ch9 を Ruby と同構成に
- domain.model（Value/List）・domain.type（Type）・application（Command）へ分割
- テストも本体 package に対応させて移動、gradle test は全 PASS
- 第 9 章のモジュール設計を Ruby 版と同じ 3 層構成に書き換え
